refactor: remove legacy config path
